Added feature to only allow users in the same session to interact with one another.
